# Improve the performance of sawyer chunk decoding