The performance was measured including my other PR #22437 but the difference shouldn't be too vast.

develop:
```
Building object index (7907 items)
Finished building object index in 7.44 seconds.
Building track design index (515 items)
Finished building track design index in 1.11 seconds.
Building scenario index (164 items)
Finished building scenario index in 0.81 seconds.
```
PR:
```
Building object index (7907 items)
Finished building object index in 0.37 seconds.
Building track design index (515 items)
Finished building track design index in 0.41 seconds.
Building scenario index (164 items)
Finished building scenario index in 0.26 seconds.
```