@tupaschoal 2024-08-03
Performance gain looks great!